pyproject.toml
[tool.poetry]
name = "sort pictures"
version = "0.1.0"
description = "This python package should be able to sort your family pictures"
authors = ["minimicha <magoussarov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^10.1.0"
pandas = "^2.1.4"
pytest = "^7.4.3"
testfixtures = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
